// api/src/constants/roles.ts
/**
 * System level roles.
 *
 * @export
 * @enum {number}
 */
export enum SYSTEM_ROLE {
  SYSTEM_ADMIN = 'System Administrator',
  DATA_ADMINISTRATOR = 'Data Administrator'
}

/**
 * Project level roles.
 *
 * @export
 * @enum {number}
 */
export enum PROJECT_ROLE {
  PROJECT_LEAD = 'Project Lead',
  PROJECT_EDITOR = 'Editor',
  PROJECT_VIEWER = 'Viewer'
}